refactor(gui): share host power menu with Diagnostics shortcut
