+++ gui/web/src/i18n/locales/en.json
@@ -1,22 +1,8 @@
 {
   "systemPower": {
-    "cancel": "Cancel",
     "title": "Host",
     "description": "Restart or shut down the computer running the mower and web interface.",
-    "reboot": "Reboot Host",
-    "shutdown": "Shut down Host",
-    "rebootTitle": "Reboot the host?",
-    "shutdownTitle": "Shut down the host?",
-    "rebootConfirm": "This restarts the whole host and all mower services. The web interface will disconnect while it restarts.",
-    "shutdownConfirm": "This shuts down the whole host and all mower services. The web interface will disconnect. Physical access is required to power it back on.",
-    "parkFirst": "Park the mower and stop the blade before continuing. Do not interrupt an update or firmware flash.",
-    "rebootRequested": "Reboot requested",
-    "shutdownRequested": "Shutdown requested",
-    "rebootResult": "The host may take a minute or two to restart. Reload this page when it is back online.",
-    "shutdownResult": "The host is shutting down. Physical access is required to power it back on.",
-    "reload": "Reload page",
-    "unconfirmed": "Request could not be confirmed",
-    "unconfirmedHint": "The connection may have closed because the host is already restarting or shutting down. Check its status before trying again; this request will not be resent automatically."
+    "open": "Host power…"
   },
   "language": {
     "label": "Language",
@@ -159,14 +145,14 @@
     "restartMowgliConfirm": "This restarts the entire Mowgli stack (ROS2, GPS, LiDAR, and this web GUI) — like the mowgli-restart command. The page will disconnect briefly and reload automatically when the GUI is back.",
     "restartingStackTitle": "Restarting Mowgli…",
     "restartingStackBody": "Restarting the whole stack. This page will reconnect and reload automatically when the GUI is back.",
-    "rebootingPiTitle": "Rebooting the Raspberry Pi…",
-    "rebootingPiBody": "The Raspberry Pi is rebooting. This page will reconnect and reload automatically when it's back (this can take a minute or two).",
+    "rebootingPiTitle": "Rebooting the Host…",
+    "rebootingPiBody": "The Host is rebooting. This page will reconnect and reload automatically when it's back (this can take a minute or two).",
     "advanced": "Advanced",
     "restartBoard": "Restart the board (STM32)",
     "restartBoardConfirm": "Resets the STM32 firmware (NVIC_SystemReset). Use it to unstick a frozen board — e.g. the IMU returning NaN. The motors/blade stop during the reset (~1 s).",
-    "restartPi": "Restart the Raspberry Pi",
+    "restartPi": "Restart Host",
     "restartPiConfirm": "The system will restart. You will lose the connection temporarily.",
-    "shutdownPi": "Shut down the Raspberry Pi",
+    "shutdownPi": "Shut down Host",
     "shutdownPiConfirm": "The system will shut down. Physical access will be required to power it back on.",
     "rearm": "Re-arm",
     "rearmEmergencyTooltip": "Re-arm the emergency stop (the firmware decides whether the latch releases)",
